Fix two bugs in r-p-s.sol
We can't use a simple hashed key to select the player1, because Bob (player1)
has to show the key in cleartext on the blockchain when posting his
transaction, at which point Eve can see the key and race him to the blockchain
with a higher-gased transaction. Back to player restriction by known address.

When Bob (player1) wins by default, he should get Alice's (player0's) bet
as well as his own and the escrow.